commands: split presentation and project boundaries
