+++ src/commands/edit.rs
@@ -1,0 +1,118 @@
+// SPDX-License-Identifier: EUPL-1.2
+
+use super::{
+    Path,
+    PinType,
+    Project,
+    Result,
+    Source,
+    Unpack,
+    bail,
+    fetch,
+    pins,
+    render,
+    shorturl,
+};
+
+pub fn add(
+    name: &str,
+    url: &str,
+    pin_type: PinType,
+    unpack: Option<Unpack>,
+    dir_field: Option<&str>,
+    submodules: bool,
+    follows: &[(String, String)],
+) -> Result<()> {
+    if unpack.is_some() && pin_type != PinType::Fixed {
+        bail!("--unpack is only valid with --fixed");
+    }
+    let project = Project::discover();
+    let mut doc = project.load_pins()?;
+    if pins::has_input(&doc, name) {
+        bail!("input '{name}' already exists");
+    }
+    pins::add_input(&mut doc, name, url, &pins::AddInputOpts {
+        pin_type,
+        unpack,
+        dir: dir_field,
+        submodules,
+        follows,
+    });
+    project.save_pins(&doc)?;
+
+    let expanded = shorturl::expand(url, &pins::shorturls(&doc));
+    let fetched = match pin_type {
+        PinType::Fixed => fetch::fetch_fixed_pin(&expanded, unpack),
+        PinType::Flake | PinType::Fetch => {
+            expanded
+                .parse::<Source>()
+                .and_then(|source| fetch::fetch_pin(&source, submodules))
+        },
+    };
+    match fetched {
+        Ok((node, rev)) => {
+            let mut lk = project.load_lock()?;
+            lk.insert(name.to_owned(), node);
+            project.save_lock(&lk)?;
+            println!("added {name}  NEW -> {}", render::short(&rev));
+        },
+        Err(err) => {
+            println!("added {name} to pins.toml, but locking failed: {err:#}");
+            println!("  fix the url and run `tack update {name}`");
+        },
+    }
+    Ok(())
+}
+
+pub fn rm(name: &str) -> Result<()> {
+    let project = Project::discover();
+    let (removed_pin, removed_lock) = rm_in_dir(project.dir(), name)?;
+    if removed_pin {
+        println!("removed {name}");
+    } else if removed_lock {
+        println!("removed stale lock entry {name}");
+    }
+    Ok(())
+}
+
+pub(in crate::commands) fn rm_in_dir(dir: &Path, name: &str) -> Result<(bool, bool)> {
+    let project = Project::at(dir.to_owned());
+    let mut doc = project.load_pins()?;
+    let removed_pin = pins::remove_input(&mut doc, name);
+
+    let mut lk = project.load_lock()?;
+    let removed_lock = lk.remove(name).is_some();
+
+    if !removed_pin && !removed_lock {
+        bail!("no input '{name}'");
+    }
+
+    if removed_pin {
+        project.save_pins(&doc)?;
+    }
+    if removed_lock {
+        project.save_lock(&lk)?;
+    }
+    Ok((removed_pin, removed_lock))
+}
+
+pub fn alias(name: &str, template: Option<&str>, remove: bool) -> Result<()> {
+    let project = Project::discover();
+    let mut doc = project.load_pins()?;
+    if remove {
+        if !pins::remove_alias(&mut doc, name) {
+            bail!("no alias '{name}'");
+        }
+        project.save_pins(&doc)?;
+        println!("removed alias {name}");
+    } else {
+        let tpl = template.expect("template required");
+        if !tpl.contains("{path}") {
+            bail!("alias template must contain '{{path}}'");
+        }
+        pins::set_alias(&mut doc, name, tpl);
+        project.save_pins(&doc)?;
+        println!("alias {name} = {tpl}");
+    }
+    Ok(())
+}
